made hooks and hide add blog form and logout if nto admin
