+++ src/app/page.tsx
@@ -2,6 +2,7 @@ import AddBlogForm from "@/components/ui/add-blog-form/add-blog-form";
 import { DisplayBlogs } from "@/components/ui/display-blogs/display-blogs";
 import { getBlogs } from "@/lib/blogs-dal";
 import { Suspense } from "react";
+import { LogoutButton } from "@/components/ui/buttons/logout-button";
 
 export default async function Home() {
   const blogsPromise = getBlogs();
@@ -12,6 +13,7 @@ export default async function Home() {
         <DisplayBlogs blogsPromise={blogsPromise} />
       </Suspense>
       <AddBlogForm />
+      <LogoutButton />
     </div>
   );
 }
