+++ src/components/ui/buttons/logout-button.tsx
@@ -1,6 +1,7 @@
 'use client';
 
 import { Button } from "@/components/ui/buttons/button";
+import { useIsAdmin } from "@/lib/hooks/hooks";
 import { signOut } from 'next-auth/react';
 
 export function LogoutButton() {
@@ -8,7 +9,12 @@ export function LogoutButton() {
     await signOut({ redirectTo: '/' });
   };
 
+  const { data } = useIsAdmin();
   return (
-    <Button onClick={handleLogout} text={"Logout"} />
+    <>
+      {data &&
+        <Button onClick={handleLogout} text={"Logout"} />
+      }
+    </>
   );
 }
