+++ src/lib/hooks/hooks.ts
@@ -1,5 +1,6 @@
 import { keepPreviousData, useQuery } from '@tanstack/react-query';
-import { BlogPublishedResponse, ResponseData, Section } from '@/types/types';
+import { BlogPublishedResponse, Section } from '@/types/types';
+import { isAdmin } from '../actions/auth';
 
 const fetchBlogSections = async (blogId: number): Promise<Section[]> => {
   const response = await fetch(`/api/sections/${blogId}`)
@@ -29,5 +30,13 @@ export const useGetIsBlogPublished = (blogId: number) => {
   return useQuery({
     queryKey: ['blog', blogId],
     queryFn: () => fetchIsBlogPublished(blogId),
+  })
+}
+
+
+export const useIsAdmin = () => {
+  return useQuery({
+    queryKey: ['isAdmin'],
+    queryFn: () => isAdmin(),
   })
 }
