+++ src/lib/mutations/mutations.ts
@@ -146,3 +146,5 @@ export const useDeleteSection = (blogId: number) => {
         }
     });
 };
+
+
